pinctrl: msm: Remove explicit barriers from mmio ops where unneeded
For the vast majority of mmio operations in this driver, explicit memory
barriers aren't needed either because a data dependency between a read
and write already exists, or because of the presence of the spin locks
which execute a full memory barrier.

Removing all the unneeded explicit barriers considerably reduces
overhead for pinctrl operations, which in turn benefits things like i2c.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[yaro: adjusted for msm-5.4's changes + applied 'pinctrl: msm: Restore some barriers to prevent reordering of I/O writes']
Signed-off-by: Yaroslav Furman <yaro330@gmail.com>